Refactor NewEntry: extract Dialogs from VernWithSug
