+++ src/components/DataEntry/DataEntryTable/NewEntry/NewEntry.tsx
@@ -2,12 +2,17 @@ import { Grid, Typography } from "@material-ui/core";
 import React from "react";
 import { Translate } from "react-localize-redux";
 
+import DupFinder, {
+  DefaultParams,
+} from "../../../../goals/MergeDupGoal/DuplicateFinder/DuplicateFinder";
 import theme from "../../../../types/theme";
 import { SemanticDomain, simpleWord, Word } from "../../../../types/word";
 import Pronunciations from "../../../Pronunciations/PronunciationsComponent";
 import Recorder from "../../../Pronunciations/Recorder";
 import GlossWithSuggestions from "../GlossWithSuggestions/GlossWithSuggestions";
 import VernWithSuggestions from "../VernWithSuggestions/VernWithSuggestions";
+import SenseDialog from "./SenseDialog";
+import VernDialog from "./VernDialog";
 
 interface NewEntryProps {
   allVerns: string[];
@@ -31,6 +36,11 @@ interface NewEntryState {
   wordId?: string;
   activeGloss: string;
   audioFileURLs: string[];
+  vernOpen: boolean;
+  senseOpen: boolean;
+  suggestedVerns: string[];
+  dupVernWords: Word[];
+  selectedWord: Word;
 }
 
 function focusInput(inputRef: React.RefObject<HTMLDivElement>) {
@@ -47,6 +57,13 @@ export default class NewEntry extends React.Component<
   NewEntryProps,
   NewEntryState
 > {
+  readonly maxSuggestions = 5;
+  readonly maxLevDistance = 3; // The default 5 allows too much distance
+  suggestionFinder: DupFinder = new DupFinder({
+    ...DefaultParams,
+    maxScore: this.maxLevDistance,
+  });
+
   constructor(props: NewEntryProps) {
     super(props);
     this.state = {
@@ -54,6 +71,11 @@ export default class NewEntry extends React.Component<
       activeGloss: "",
       audioFileURLs: [],
       isDupVern: false,
+      vernOpen: false,
+      senseOpen: false,
+      suggestedVerns: [],
+      dupVernWords: [],
+      selectedWord: { ...simpleWord("", ""), id: "" },
     };
     this.vernInput = React.createRef<HTMLDivElement>();
     this.glossInput = React.createRef<HTMLDivElement>();
@@ -91,7 +113,8 @@ export default class NewEntry extends React.Component<
     this.setState({ newEntry, activeGloss: newValue });
   }
 
-  updateVernField(newValue: string): Word[] {
+  updateVernField(newValue: string, openDialog?: boolean) {
+    this.props.setIsReadyState(newValue.trim().length > 0);
     let dupVernWords: Word[] = [];
     let isDupVern: boolean = false;
     if (newValue) {
@@ -105,7 +128,12 @@ export default class NewEntry extends React.Component<
     }
     const newEntry = { ...this.state.newEntry, vernacular: newValue };
     this.setState({ isDupVern, newEntry });
-    return dupVernWords;
+    this.updateSuggestedVerns(newValue);
+    if (openDialog && isDupVern) {
+      this.setState({ vernOpen: true, dupVernWords });
+    } else {
+      this.updateWordId();
+    }
   }
 
   updateWordId(wordId?: string) {
@@ -163,7 +191,7 @@ export default class NewEntry extends React.Component<
   }
 
   async handleEnterAndTab(e: React.KeyboardEvent) {
-    if (e.key === "Enter") {
+    if (!this.state.vernOpen && e.key === "Enter") {
       if (this.state.newEntry.vernacular) {
         if (this.state.activeGloss) {
           await this.addOrUpdateWord();
@@ -176,6 +204,95 @@ export default class NewEntry extends React.Component<
         this.focusVernInput();
       }
     }
+  }
+
+  openDialog() {
+    this.setState({ vernOpen: true });
+  }
+
+  handleCloseVernDialog(selectedWordId?: string) {
+    this.setState({ vernOpen: false }, () => {
+      this.updateWordId(selectedWordId);
+    });
+    if (selectedWordId) {
+      let selectedWord: Word = this.state.dupVernWords.find(
+        (word: Word) => word.id === selectedWordId
+      )!;
+      this.setState({
+        selectedWord,
+        senseOpen: true,
+      });
+    } else if (selectedWordId === "") {
+      let selectedWord: Word = {
+        ...simpleWord(this.state.newEntry.vernacular, ""),
+        id: "",
+      };
+      this.setState({ selectedWord });
+    }
+  }
+
+  handleCloseSenseDialog(senseIndex?: number) {
+    if (senseIndex === undefined) {
+      this.updateWordId();
+      this.setState({
+        selectedWord: { ...simpleWord("", ""), id: "" },
+        vernOpen: true,
+      });
+    } else if (senseIndex >= 0) {
+      const activeGloss = this.state.selectedWord.senses[senseIndex].glosses[0]
+        .def;
+      this.setState({ activeGloss });
+    }
+    this.setState({ senseOpen: false });
+  }
+
+  autoCompleteCandidates(vernacular: string): string[] {
+    // filter allVerns to those that start with vernacular
+    // then map them into an array sorted by length and take the 2 shortest
+    // and the rest longest (should make finding the long words easier)
+    let scoredStartsWith: [string, number][] = [];
+    let startsWith = this.props.allVerns.filter((vern: string) =>
+      vern.startsWith(vernacular)
+    );
+    for (const v of startsWith) {
+      scoredStartsWith.push([v, v.length]);
+    }
+    let keepers = scoredStartsWith
+      .sort((a, b) => a[1] - b[1])
+      .map((vern) => vern[0]);
+    if (keepers.length > this.maxSuggestions) {
+      keepers.splice(2, keepers.length - this.maxSuggestions);
+    }
+    return keepers;
+  }
+
+  updateSuggestedVerns(value?: string | null) {
+    let suggestedVerns: string[] = [];
+    if (value) {
+      suggestedVerns = [...this.autoCompleteCandidates(value)];
+      if (suggestedVerns.length < this.maxSuggestions) {
+        const viableVerns: string[] = this.props.allVerns.filter(
+          (vern: string) =>
+            this.suggestionFinder.getLevenshteinDistance(vern, value) <
+            this.suggestionFinder.maxScore
+        );
+        const sortedVerns: string[] = viableVerns.sort(
+          (a: string, b: string) =>
+            this.suggestionFinder.getLevenshteinDistance(a, value) -
+            this.suggestionFinder.getLevenshteinDistance(b, value)
+        );
+        let candidate: string;
+        while (
+          suggestedVerns.length < this.maxSuggestions &&
+          sortedVerns.length
+        ) {
+          candidate = sortedVerns.shift()!;
+          if (!suggestedVerns.includes(candidate))
+            suggestedVerns.push(candidate);
+        }
+      }
+    }
+    this.setState({ suggestedVerns });
   }
 
   render() {
@@ -197,18 +314,33 @@ export default class NewEntry extends React.Component<
                 isNew={true}
                 vernacular={this.state.newEntry.vernacular}
                 vernInput={this.vernInput}
-                updateVernField={(newValue: string) => {
-                  this.props.setIsReadyState(newValue.trim().length > 0);
-                  return this.updateVernField(newValue);
+                updateVernField={(newValue: string, openDialog?: boolean) => {
+                  this.updateVernField(newValue, openDialog);
                 }}
-                updateWordId={(wordId?: string) => this.updateWordId(wordId)}
-                selectedWordId={this.state.wordId}
-                allVerns={this.props.allVerns}
+                onBlur={() => {
+                  this.updateVernField(
+                    this.state.newEntry.vernacular,
+                    this.state.wordId === undefined && this.state.isDupVern
+                  );
+                }}
+                suggestedVerns={this.state.suggestedVerns}
                 handleEnterAndTab={(e: React.KeyboardEvent) =>
                   this.handleEnterAndTab(e)
                 }
-                setActiveGloss={(newGloss: string) =>
-                  this.setState({ activeGloss: newGloss })
+              />
+              <VernDialog
+                open={this.state.vernOpen}
+                handleClose={(selectedWordId?: string) =>
+                  this.handleCloseVernDialog(selectedWordId)
+                }
+                vernacularWords={this.state.dupVernWords}
+                analysisLang={this.props.analysisLang}
+              />
+              <SenseDialog
+                selectedWord={this.state.selectedWord}
+                open={this.state.senseOpen}
+                handleClose={(senseIndex?: number) =>
+                  this.handleCloseSenseDialog(senseIndex)
                 }
                 analysisLang={this.props.analysisLang}
               />
