+++ src/components/DataEntry/DataEntryTable/VernWithSuggestions/VernWithSuggestions.tsx
@@ -7,119 +7,23 @@ import {
   withLocalize,
 } from "react-localize-redux";
 
-import { simpleWord, Word } from "../../../../types/word";
-import DupFinder, {
-  DefaultParams,
-} from "../../../../goals/MergeDupGoal/DuplicateFinder/DuplicateFinder";
-import SenseDialog from "./SenseDialog/SenseDialog";
-import VernDialog from "./VernDialog/VernDialog";
-
 interface VernWithSuggestionsProps {
   isNew?: boolean;
   isDisabled?: boolean;
   vernacular: string;
   vernInput?: React.RefObject<HTMLDivElement>;
-  updateVernField: (newValue: string) => Word[];
-  setActiveGloss: (newGloss: string) => void;
-  allVerns: string[];
+  updateVernField: (newValue: string, openDialog?: boolean) => void;
+  onBlur: () => void;
+  suggestedVerns?: string[];
   handleEnterAndTab: (e: React.KeyboardEvent) => void;
-  updateWordId: (wordId?: string) => void;
-  selectedWordId?: string;
-  onBlur?: () => void;
-  analysisLang: string;
-}
-
-interface VernWithSuggestionsState {
-  vernOpen: boolean;
-  senseOpen: boolean;
-  suggestedVerns: string[];
-  dupVernWords: Word[];
-  selectedWord: Word;
 }
 
 /**
  * An editable vernacular field for new words
  */
 export class VernWithSuggestions extends React.Component<
-  LocalizeContextProps & VernWithSuggestionsProps,
-  VernWithSuggestionsState
+  LocalizeContextProps & VernWithSuggestionsProps
 > {
-  readonly maxSuggestions = 5;
-  readonly maxLevDistance = 3; // The default 5 allows too much distance
-  suggestionFinder: DupFinder = new DupFinder({
-    ...DefaultParams,
-    maxScore: this.maxLevDistance,
-  });
-
-  constructor(props: any) {
-    super(props);
-    this.state = {
-      vernOpen: false,
-      senseOpen: false,
-      suggestedVerns: [],
-      dupVernWords: [],
-      selectedWord: { ...simpleWord("", ""), id: "" },
-    };
-  }
-
-  autoCompleteCandidates(vernacular: string): string[] {
-    // filter allVerns to those that start with vernacular
-    // then map them into an array sorted by length and take the 2 shortest
-    // and the rest longest (should make finding the long words easier)
-    let scoredStartsWith: [string, number][] = [];
-    let startsWith = this.props.allVerns.filter((vern: string) =>
-      vern.startsWith(vernacular)
-    );
-    for (const v of startsWith) {
-      scoredStartsWith.push([v, v.length]);
-    }
-    let keepers = scoredStartsWith
-      .sort((a, b) => a[1] - b[1])
-      .map((vern) => vern[0]);
-    if (keepers.length > this.maxSuggestions) {
-      keepers.splice(2, keepers.length - this.maxSuggestions);
-    }
-    return keepers;
-  }
-
-  updateSuggestedVerns(value?: string | null) {
-    let suggestedVerns: string[] = [];
-    if (value) {
-      suggestedVerns = [...this.autoCompleteCandidates(value)];
-      if (suggestedVerns.length < this.maxSuggestions) {
-        const viableVerns: string[] = this.props.allVerns.filter(
-          (vern: string) =>
-            this.suggestionFinder.getLevenshteinDistance(vern, value) <
-            this.suggestionFinder.maxScore
-        );
-        const sortedVerns: string[] = viableVerns.sort(
-          (a: string, b: string) =>
-            this.suggestionFinder.getLevenshteinDistance(a, value) -
-            this.suggestionFinder.getLevenshteinDistance(b, value)
-        );
-        let candidate: string;
-        while (
-          suggestedVerns.length < this.maxSuggestions &&
-          sortedVerns.length
-        ) {
-          candidate = sortedVerns.shift()!;
-          if (!suggestedVerns.includes(candidate))
-            suggestedVerns.push(candidate);
-        }
-      }
-    }
-    this.setState({ suggestedVerns });
-  }
-
-  handleSelection(value: string) {
-    let dupVernWords: Word[] = this.props.updateVernField(value);
-    if (dupVernWords.length > 0) {
-      this.setState({ vernOpen: true, dupVernWords });
-    } else {
-      this.props.updateWordId();
-    }
-  }
-
   render() {
     return (
       <React.Fragment>
@@ -127,30 +31,23 @@ export class VernWithSuggestions extends React.Component<
           freeSolo
           disabled={this.props.isDisabled}
           value={this.props.vernacular}
-          options={this.state.suggestedVerns}
-          onBlur={(_event: React.FocusEvent<HTMLDivElement>) => {
-            if (this.props.onBlur) {
-              this.props.onBlur();
-            }
-            if (this.props.selectedWordId === undefined)
-              this.handleSelection(this.props.vernacular);
-          }}
-          onChange={(_event, value) => {
+          options={this.props.suggestedVerns ? this.props.suggestedVerns : []}
+          onBlur={this.props.onBlur}
+          onChange={(_e, value) => {
             // onChange is triggered when an option is selected
-            if (!value) value = "";
-            this.handleSelection(value);
-            this.updateSuggestedVerns(value);
+            if (!value) {
+              value = "";
+            }
+            this.props.updateVernField(value, true);
           }}
           onInputChange={(_event, value) => {
             // onInputChange is triggered by typing
-            const dupVernWords = this.props.updateVernField(value);
-            this.setState({ dupVernWords });
-            this.updateSuggestedVerns(value);
-            this.props.updateWordId();
+            this.props.updateVernField(value);
           }}
           onKeyPress={(e: React.KeyboardEvent) => {
-            if (!this.state.vernOpen && (e.key === "Enter" || e.key === "Tab"))
+            if (e.key === "Enter" || e.key === "Tab") {
               this.props.handleEnterAndTab(e);
+            }
           }}
           renderInput={(params) => (
             <TextField
@@ -164,54 +61,6 @@ export class VernWithSuggestions extends React.Component<
             />
           )}
         />
-        {this.props.isNew && (
-          <VernDialog
-            open={this.state.vernOpen}
-            handleClose={(selectedWordId?: string) => {
-              this.setState({ vernOpen: false }, () => {
-                this.props.updateWordId(selectedWordId);
-              });
-              if (selectedWordId) {
-                let selectedWord: Word = this.state.dupVernWords.find(
-                  (word: Word) => word.id === selectedWordId
-                )!;
-                this.setState({
-                  selectedWord,
-                  senseOpen: true,
-                });
-              } else if (selectedWordId === "") {
-                let selectedWord: Word = {
-                  ...simpleWord(this.props.vernacular, ""),
-                  id: "",
-                };
-                this.setState({ selectedWord });
-              }
-            }}
-            vernacularWords={this.state.dupVernWords}
-            analysisLang={this.props.analysisLang}
-          />
-        )}
-        {this.props.isNew && (
-          <SenseDialog
-            selectedWord={this.state.selectedWord}
-            open={this.state.senseOpen}
-            handleClose={(senseIndex?: number) => {
-              if (senseIndex === undefined) {
-                this.props.updateWordId();
-                this.setState({
-                  selectedWord: { ...simpleWord("", ""), id: "" },
-                  vernOpen: true,
-                });
-              } else if (senseIndex >= 0) {
-                this.props.setActiveGloss(
-                  this.state.selectedWord.senses[senseIndex].glosses[0].def
-                );
-              }
-              this.setState({ senseOpen: false });
-            }}
-            analysisLang={this.props.analysisLang}
-          />
-        )}
       </React.Fragment>
     );
   }
